Added Flexibility for in_path Argument to optionally be an existing data frame. Moved mair_ls_serial to msatutil/mair_ls.py, where it better belongs.
